new service and some html
added a new service controller for the second action we want to implenet about the traffic in every toll for a company.
Created 2 html files. The first one shows the action of tollStationPasses and the second one shows the activity in each toll of a company with a specific tollOpID

NOTE!!!!

you have to update the toke before you run the code. TODO fix it